Wrap links and logo into anchor tags and wrap the header and description into a single div called top
This will make it so that the links and logo will
take you to a different webpage. In this case the
links won't take you anywhere and the logo will
direct the user to the home page. The additional
div was added because the header and description
both shared the same background color.