Add a div around the images in the gallery and add a new div around the call to action quote
The div around the images allows the blue border
to be styled around it and to make sure the
width and height are correct. The div around the
call to action quote makes sure that the entire
quote is treated as a single flex item in it's
repective flex container.